requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
pandas
google-cloud-sql-connector
sqlalchemy
pg8000